% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/av_fun_GVC.R
\name{Agreg_Manuf}
\alias{Agreg_Manuf}
\title{Agreg_Manuf
greggation function
TypAgreg= Manuf_CT ;	Manuf_IP19 ;	Manuf_CT_lrwiod ;	Manuf_IP19_lrwiod     MRIO= LRWIOD ; WIOD; FIGARO
Warning : the variable to agreggate must be named "value"
Use of StructDocs to manage classifications}
\usage{
Agreg_Manuf(DT, MRIO, Var_To_Agreg, TypAgreg)
}
\arguments{
\item{DT}{datatable}

\item{MRIO}{MRIO object}

\item{Var_To_Agreg}{text variable}

\item{TypAgreg}{text}
}
\value{
dl data long
}
\description{
Agreg_Manuf
greggation function
TypAgreg= Manuf_CT ;	Manuf_IP19 ;	Manuf_CT_lrwiod ;	Manuf_IP19_lrwiod     MRIO= LRWIOD ; WIOD; FIGARO
Warning : the variable to agreggate must be named "value"
Use of StructDocs to manage classifications
}
