% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/av_heart_MRIO_manager.R
\name{AjoutPRBR}
\alias{AjoutPRBR}
\title{AjoutPRBR
Add PR and BR to help transposing tables to wide format (eg. matrix like A or L)}
\usage{
AjoutPRBR(dl)
}
\arguments{
\item{dl}{long data}
}
\value{
dl data long
}
\description{
AjoutPRBR
Add PR and BR to help transposing tables to wide format (eg. matrix like A or L)
}
\examples{
\dontrun{ f_dt <- AjoutPRBR(interm)
y_dt <- AjoutPRBR(y_dt)}
}
