% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/av_fun_GVC.R
\name{Attr_TxSimu_HRM_100MoE}
\alias{Attr_TxSimu_HRM_100MoE}
\title{Attr_TxSimu_HRM_100MoE
Function to affect the right % corresponding to 100Mo of VA or PROD.
Why 100Mo ? / because as we exchange imports with production we need to have enough imports. --> Linear Model => results*10 with no problem.
TypeTx = "VA" or "PROD"}
\usage{
Attr_TxSimu_HRM_100MoE(
  Produit,
  pays,
  annee,
  TypeTx = "VA",
  OptSourceRDS = "XXXXX",
  OptMRIOlong = NULL
)
}
\arguments{
\item{Produit}{text industry}

\item{pays}{text country}

\item{annee}{year}

\item{TypeTx}{text}

\item{OptSourceRDS}{binary}

\item{OptMRIOlong}{binary}
}
\value{
numeric value
}
\description{
Attr_TxSimu_HRM_100MoE
Function to affect the right % corresponding to 100Mo of VA or PROD.
Why 100Mo ? / because as we exchange imports with production we need to have enough imports. --> Linear Model => results*10 with no problem.
TypeTx = "VA" or "PROD"
}
