% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/av_fun_GVC.R
\name{Autarky}
\alias{Autarky}
\title{Autarky
Function for calculating an autarky situation from a MRIO which is the starting world economy situation
The MRIO has already undergone a CompoMRIO, and the Save option does a saveRDS in "MRIO_Autarky.rds".}
\usage{
Autarky(dtdl, Optdl = FALSE, OptSaveRDS = FALSE, OptBaseDT = FALSE)
}
\arguments{
\item{dtdl}{datatable}

\item{Optdl}{binary}

\item{OptSaveRDS}{binary}

\item{OptBaseDT}{binary}
}
\value{
dl data long
}
\description{
Autarky
Function for calculating an autarky situation from a MRIO which is the starting world economy situation
The MRIO has already undergone a CompoMRIO, and the Save option does a saveRDS in "MRIO_Autarky.rds".
}
