% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/av_create_MRIO_LF.R
\encoding{UTF-8}
\name{av_create_FIGAROixi_2022}
\alias{av_create_FIGAROixi_2022}
\title{av_create_FIGAROixi_2022}
\usage{
av_create_FIGAROixi_2022(Path_FIG22ixi, Path_out, OptAnnual = FALSE)
}
\arguments{
\item{Path_FIG22ixi}{path of FIGARO ixi CSV flat format files}

\item{Path_out}{path to save normalized data}

\item{OptAnnual}{option to keep one files by year instead of one file with all years. Can provide a degraded solution if a single database requires too much memory.}
}
\value{
Nothing. Only save data into normalized format.
}
\description{
Figaro 2022 release, CSV flat format : link : https://ec.europa.eu/eurostat/web/esa-supply-use-input-tables/database#CSV%20flat%20format%20(FIGARO%202022%20edition)
}
\examples{
\dontrun{
Path_FIG22ixi <- paste0(PathTemp, "ixi/")
Path_out <- PathTest
av_create_FIGAROixi_2022(Path_FIG22ixi, Path_out, OptAnnual = TRUE)
av_create_FIGAROixi_2022(Path_FIG22ixi, Path_out, OptAnnual = FALSE)
testAll <- readRDS(paste0(Path_out, "/BDn_FIG.rds"))
test2015 <- readRDS(paste0(Path_out, "/BDn_FIG_", 2015, ".rds"))}
}
