% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/av_create_MRIO_LF.R
\encoding{UTF-8}
\name{av_create_LRWIOD_2022}
\alias{av_create_LRWIOD_2022}
\title{av_create_LRWIOD_2022
LR-WIOD 2022 release : link : https://www.rug.nl/ggdc/valuechain/long-run-wiod?lang=en
License and funding : Long-run WIOD is licensed under a Creative Commons Attribution 4.0 International License . The construction of the Long-run WIOD was supported by the Dutch Science Foundation (NWO) [grant number 453-14-012].}
\usage{
av_create_LRWIOD_2022(Path_WIODLR, Path_out, ConvertToEuro = TRUE)
}
\arguments{
\item{Path_WIODLR}{path of Long-run WIOD flat CSV format files}

\item{Path_out}{path to save normalized data}

\item{ConvertToEuro}{Option to convert Dollars to Euros}
}
\value{
Nothing. Only save data into normalized format.
}
\description{
av_create_LRWIOD_2022
LR-WIOD 2022 release : link : https://www.rug.nl/ggdc/valuechain/long-run-wiod?lang=en
License and funding : Long-run WIOD is licensed under a Creative Commons Attribution 4.0 International License . The construction of the Long-run WIOD was supported by the Dutch Science Foundation (NWO) [grant number 453-14-012].
}
\examples{
\dontrun{ Path_WIODLR <- PathTemp
Path_out <- PathTest
av_create_LRWIOD_2022(Path_WIODLR, Path_out, ConvertToEuro = TRUE)
testEuroDolls <- readRDS(paste0(Path_out, "/BDn_LR_WIOD.rds"))
Euros <- testEuroDolls
av_create_LRWIOD_2022(Path_WIODLR, Path_out, ConvertToEuro = FALSE)
testEuroDolls <- readRDS(paste0(Path_out, "/BDn_LR_WIOD.rds"))
dolls <- testEuroDolls}
}
