% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/av_create_MRIO_LF.R
\name{av_dl_MicroMRIO}
\alias{av_dl_MicroMRIO}
\title{av_dl_MicroMRIO}
\usage{
av_dl_MicroMRIO(dl, OptXLSout = FALSE)
}
\arguments{
\item{dl}{long data}

\item{OptXLSout}{to export the mini-MRIO in Excel format}
}
\value{
dl micro
}
\description{
av_dl_MicroMRIO
}
\examples{
\dontrun{ DT_micro <- av_dl_MicroMRIO(DT, OptXLSout = FALSE)
DT_microXls <- av_dl_MicroMRIO(DT, OptXLSout = TRUE)}
}
