% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/av_create_MRIO_LF.R
\name{av_dl_UE27}
\alias{av_dl_UE27}
\title{av_dl_UE27
Build a new MRIO with all EU countries aggragated into "UE27" item.
Useful to calculate EU made-in for instance.
WARNING : dl must be normalized (not raw data) for countries homogeneity reasons.}
\usage{
av_dl_UE27(dl, OptSaveRDS = "NO")
}
\arguments{
\item{dl}{long data}

\item{OptSaveRDS}{text}
}
\value{
dl with EU27 agreggate instead of each country
}
\description{
av_dl_UE27
Build a new MRIO with all EU countries aggragated into "UE27" item.
Useful to calculate EU made-in for instance.
WARNING : dl must be normalized (not raw data) for countries homogeneity reasons.
}
\examples{
\dontrun{ DT_UE27 <- av_dl_UE27(DT)}
}
