% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/av_extensions_manager.R
\name{av_extend_MRIO_dw}
\alias{av_extend_MRIO_dw}
\title{av_extend_MRIO_dw
Extend MRIO components by adding stressors like Value added or CO2 Emissions
You can manage this function to add new cases : other MRIO options or other stressors etc.}
\usage{
av_extend_MRIO_dw(MRIO_dw, NameMRIO, TypExtension, Path1 = NULL)
}
\arguments{
\item{MRIO_dw}{MRIO must be in wide format (not Long)}

\item{NameMRIO}{text}

\item{TypExtension}{text}

\item{Path1}{text}
}
\value{
dw data wide (list)
}
\description{
av_extend_MRIO_dw
Extend MRIO components by adding stressors like Value added or CO2 Emissions
You can manage this function to add new cases : other MRIO options or other stressors etc.
}
\examples{
\dontrun{ Bonus <- CompoMRIO(DT, typeCompo = "OptFullOptionsBonus", OptTab = FALSE)
MRIO2 <- av_extend_MRIO_dw(MRIO_dw = Bonus, "FIGARO", TypExtension = "StressVA")
MRIO3 <- av_extend_MRIO_dw(MRIO_dw = Bonus, "FIGARO", TypExtension = "StressEmi", Path1 = PathTemp)
MRIO4 <- av_extend_MRIO_dw(MRIO3, "FIGARO", TypExtension = "StressVA")}
}
