% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/av_fun_GVC.R
\name{av_fun_VarPostTransition}
\alias{av_fun_VarPostTransition}
\title{av_fun_VarPostTransition
Function for robustness checks by applying structural changes at the wolrd economy to suit climate change
Parameters description :
- paramEmploi : Employment intensity (value = 0.9 => 10% drop in the employment intensity of the automotive industry)
- paramA : Intermediate consumption substitution intensity: value = 0.9 => the technical coefficient of each country's automotive industry (C29) in products C29 and C28 (domestic and imported) is reduced by 10%, and the sum of these two amounts is added to intermediate consumption of product C29 in product C27, for each country.
- paramDF : The intensity of final consumption substitution: value = 0.9 => each country's final demand for oil products (C19) is reduced by 10%, and this amount is transferred to electricity products (D35) for each country.
The adjustments made are therefore zero-sum on intermediate consumption and final demand. The simulations are called by their X-Y-Z parameters: for example, 09-08-07 will correspond to a parameter of 0.9 for employment, 0.8 for intermediate consumption and 0.7 for final demand.}
\usage{
av_fun_VarPostTransition(paramEmploi, paramA, paramDF, OptRDS = FALSE)
}
\arguments{
\item{paramEmploi}{value percentage}

\item{paramA}{value percentage}

\item{paramDF}{value percentage}

\item{OptRDS}{binary}
}
\value{
comparison table
}
\description{
av_fun_VarPostTransition
Function for robustness checks by applying structural changes at the wolrd economy to suit climate change
Parameters description :
- paramEmploi : Employment intensity (value = 0.9 => 10% drop in the employment intensity of the automotive industry)
- paramA : Intermediate consumption substitution intensity: value = 0.9 => the technical coefficient of each country's automotive industry (C29) in products C29 and C28 (domestic and imported) is reduced by 10%, and the sum of these two amounts is added to intermediate consumption of product C29 in product C27, for each country.
- paramDF : The intensity of final consumption substitution: value = 0.9 => each country's final demand for oil products (C19) is reduced by 10%, and this amount is transferred to electricity products (D35) for each country.
The adjustments made are therefore zero-sum on intermediate consumption and final demand. The simulations are called by their X-Y-Z parameters: for example, 09-08-07 will correspond to a parameter of 0.9 for employment, 0.8 for intermediate consumption and 0.7 for final demand.
}
\examples{
\dontrun{ av_fun_VarPostTransition(paramEmploi = 1, paramA = 1, paramDF = 1)
av_fun_VarPostTransition(paramEmploi = 0.9, paramA = 0.8, paramDF = 0.7)}
}
