% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/av_fun_GVC.R
\name{av_MadeIn}
\alias{av_MadeIn}
\title{av_MadeIn
Function calculation of MADE-IN}
\usage{
av_MadeIn(
  dtdl,
  Optdl = TRUE,
  annee,
  OptDonneesBrutes = FALSE,
  MadeInPays = "FRA",
  OptUE27 = FALSE
)
}
\arguments{
\item{dtdl}{data}

\item{Optdl}{binary}

\item{annee}{year}

\item{OptDonneesBrutes}{binary}

\item{MadeInPays}{text country}

\item{OptUE27}{binary}
}
\value{
dl data long
}
\description{
av_MadeIn
Function calculation of MADE-IN
}
