% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/av_heart_MRIO_manager.R
\name{av_MRIO_comparison}
\alias{av_MRIO_comparison}
\title{av_MRIO_comparison
Function to compare 2 MRIOs
MRIOs can be initially in long format (Optdl) or in wide format (split comparison of all components of the MRIO)}
\usage{
av_MRIO_comparison(MRIO1, MRIO2, Optdl = TRUE, OptVerbose = FALSE)
}
\arguments{
\item{MRIO1}{MRIO object}

\item{MRIO2}{MRIO object}

\item{Optdl}{long data option}

\item{OptVerbose}{verbose option}
}
\value{
An ordered comparison (if dl) or a list of ordered comparisons (if wide format)
}
\description{
av_MRIO_comparison
Function to compare 2 MRIOs
MRIOs can be initially in long format (Optdl) or in wide format (split comparison of all components of the MRIO)
}
\examples{
\dontrun{ M_test <- av_MRIO_comparison(Bonus, BonusBis, Optdl = F, OptVerbose = T)}
}
