% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/av_fun_GVC.R
\name{BoucleAnneesMADEINs}
\alias{BoucleAnneesMADEINs}
\title{BoucleAnneesMADEINs
Loop to build insight database on MADE-IN
2 options : (1) On the full MRIO database (if memory allows) (2) by year but need to load all the dl and put them into a list (ListdlAnnual)}
\usage{
BoucleAnneesMADEINs(
  dtdl,
  Optdl = TRUE,
  period,
  OptDonneesBrutes = FALSE,
  OptAnnual = FALSE,
  ListdlAnnual = list(),
  OptUE27 = FALSE
)
}
\arguments{
\item{dtdl}{text}

\item{Optdl}{binary}

\item{period}{time period}

\item{OptDonneesBrutes}{binary}

\item{OptAnnual}{binary}

\item{ListdlAnnual}{list}

\item{OptUE27}{binary}
}
\value{
dl data long
}
\description{
BoucleAnneesMADEINs
Loop to build insight database on MADE-IN
2 options : (1) On the full MRIO database (if memory allows) (2) by year but need to load all the dl and put them into a list (ListdlAnnual)
}
