% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/av_fun_GVC.R
\name{BoucleLinkageBwdFwd}
\alias{BoucleLinkageBwdFwd}
\title{BoucleLinkageBwdFwd
Function loop to calculate Backward and Forward Linkage indicators databases}
\usage{
BoucleLinkageBwdFwd(base_dt, period)
}
\arguments{
\item{base_dt}{datatable}

\item{period}{time period}
}
\value{
dl data long
}
\description{
BoucleLinkageBwdFwd
Function loop to calculate Backward and Forward Linkage indicators databases
}
