% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/av_fun_GVC.R
\name{BouclePaysContVAdesExports}
\alias{BouclePaysContVAdesExports}
\title{BouclePaysContVAdesExports
Calculation loop of content of value addend embedded in exports for a list of countries and a given year
Foster-McGregor, N., et R. Stehrer (2013) : “Value added content of trade : A comprehensive approach,” Economics Letters, 120(2), 354–357.}
\usage{
BouclePaysContVAdesExports(
  dtdl,
  Optdl = TRUE,
  annee,
  ListCountries = list(),
  OptUE27 = FALSE
)
}
\arguments{
\item{dtdl}{data}

\item{Optdl}{binary}

\item{annee}{year}

\item{ListCountries}{list}

\item{OptUE27}{binary}
}
\value{
dl data long
}
\description{
BouclePaysContVAdesExports
Calculation loop of content of value addend embedded in exports for a list of countries and a given year
Foster-McGregor, N., et R. Stehrer (2013) : “Value added content of trade : A comprehensive approach,” Economics Letters, 120(2), 354–357.
}
