% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/av_fun_GVC.R
\name{BouclePaysEtAnneesContVAdesExports}
\alias{BouclePaysEtAnneesContVAdesExports}
\title{BouclePaysEtAnneesContVAdesExports
Calculation loop of content of value addend embedded in exports for a list of countries and a time period (several years)
if OptAnnual=TRUE then dtdl must be a list of dtdl
Foster-McGregor, N., et R. Stehrer (2013) : “Value added content of trade : A comprehensive approach,” Economics Letters, 120(2), 354–357.}
\usage{
BouclePaysEtAnneesContVAdesExports(
  dtdl,
  Optdl = TRUE,
  period,
  ListCountries = list(),
  OptUE27 = FALSE,
  OptAnnual = FALSE
)
}
\arguments{
\item{dtdl}{data}

\item{Optdl}{binary}

\item{period}{time period}

\item{ListCountries}{list}

\item{OptUE27}{binary}

\item{OptAnnual}{binary}
}
\value{
dl data long
}
\description{
BouclePaysEtAnneesContVAdesExports
Calculation loop of content of value addend embedded in exports for a list of countries and a time period (several years)
if OptAnnual=TRUE then dtdl must be a list of dtdl
Foster-McGregor, N., et R. Stehrer (2013) : “Value added content of trade : A comprehensive approach,” Economics Letters, 120(2), 354–357.
}
