% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/av_fun_GVC.R
\name{Build_MadeIn_byOrigin}
\alias{Build_MadeIn_byOrigin}
\title{Build_MadeIn_byOrigin
 Build contents of VA by country of origin}
\usage{
Build_MadeIn_byOrigin(
  BDn_LR_WIOD,
  BDn_WIOD,
  BDn_FIG,
  BDn_ICIO,
  SelectCountry = "FRA",
  OptSaveRDS = FALSE
)
}
\arguments{
\item{BDn_LR_WIOD}{datatable}

\item{BDn_WIOD}{datatable}

\item{BDn_FIG}{datatable}

\item{BDn_ICIO}{datatable}

\item{SelectCountry}{text country}

\item{OptSaveRDS}{binary}
}
\value{
dl data long
}
\description{
Build_MadeIn_byOrigin
 Build contents of VA by country of origin
}
