% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{CFPcalculationRCPP}
\alias{CFPcalculationRCPP}
\title{Matrix Multiplication : 3 components}
\usage{
CFPcalculationRCPP(ef, L, FD)
}
\arguments{
\item{ef}{matrix}

\item{L}{matrix}

\item{FD}{matrix}
}
\description{
This function returns the matrix multiplication of 3 matrix
as for footprint calculation
}
