% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/av_heart_MRIO_manager.R
\name{CompoMRIO}
\alias{CompoMRIO}
\title{CompoMRIO
Master function in the ecosystem
Allow to build Wide MRIO (dw) from Long MRIO (dl), by splitting components
You can add other components if needed
Afterward you can use av_extend_MRIO_dw() function to extend with environmental data for example}
\usage{
CompoMRIO(MRIO_dl, typeCompo, date = 9999, OptTab = FALSE, OptUE27 = FALSE)
}
\arguments{
\item{MRIO_dl}{long format MRIO}

\item{typeCompo}{option type of composition}

\item{date}{year}

\item{OptTab}{tabular output option}

\item{OptUE27}{option with UE27 modality}
}
\value{
dw data wide (list of components)
}
\description{
CompoMRIO
Master function in the ecosystem
Allow to build Wide MRIO (dw) from Long MRIO (dl), by splitting components
You can add other components if needed
Afterward you can use av_extend_MRIO_dw() function to extend with environmental data for example
}
\examples{
\dontrun{ DT <- readRDS(paste0(PathTest, "BDn_FIG_2010.rds"))
Bonus <- CompoMRIO(DT, typeCompo = "OptFullOptionsBonus", date = 2010, OptTab = FALSE)}
}
