% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/av_fun_GVC.R
\name{ContentVAExports_Retropolation}
\alias{ContentVAExports_Retropolation}
\title{ContentVAExports_Retropolation
Retropolation of content in value added of exports WIOD is the reference and LRWIOD is used to retropolate backward, and FIGARO is used to retropolate forward.
ICIO is not used here because of systematic retropolation method
Very specific task : can be adapted depending on data you're using}
\usage{
ContentVAExports_Retropolation(
  ContVAdesExports_LRWIOD,
  ContVAdesExports_WIOD,
  ContVAdesExports_FIGARO,
  OptSaveRDS = FALSE
)
}
\arguments{
\item{ContVAdesExports_LRWIOD}{datatable}

\item{ContVAdesExports_WIOD}{datatable}

\item{ContVAdesExports_FIGARO}{datatable}

\item{OptSaveRDS}{binary}
}
\value{
dl data long
}
\description{
ContentVAExports_Retropolation
Retropolation of content in value added of exports WIOD is the reference and LRWIOD is used to retropolate backward, and FIGARO is used to retropolate forward.
ICIO is not used here because of systematic retropolation method
Very specific task : can be adapted depending on data you're using
}
