% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/av_fun_GVC.R
\name{ContentVAExports_Retropolation_UE27}
\alias{ContentVAExports_Retropolation_UE27}
\title{ContentVAExports_Retropolation_UE27
Retropolation of content in value added of exports for UE27 and big areas
Need to aggregate data before to have only one area EU
Very specific task : can be adapted depending on data you're using}
\usage{
ContentVAExports_Retropolation_UE27(
  resMadeInUE27_LRWIOD,
  resMadeInUE27_WIOD,
  resMadeInUE27_FIGARO,
  resMadeInUE27_ICIO,
  OptGraph = FALSE,
  OptSaveRDS = FALSE
)
}
\arguments{
\item{resMadeInUE27_LRWIOD}{datatable}

\item{resMadeInUE27_WIOD}{datatable}

\item{resMadeInUE27_FIGARO}{datatable}

\item{resMadeInUE27_ICIO}{datatable}

\item{OptGraph}{binary}

\item{OptSaveRDS}{binary}
}
\value{
dl data long
}
\description{
ContentVAExports_Retropolation_UE27
Retropolation of content in value added of exports for UE27 and big areas
Need to aggregate data before to have only one area EU
Very specific task : can be adapted depending on data you're using
}
