% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/av_fun_GVC.R
\name{Contenus}
\alias{Contenus}
\title{Contenus
Master function in the ecosystem
Function calculation of contents (in VA, émissions CO2, emploi) : allow footprint calculation along different formats
WARNING Needs CompoMRIO with bonus and extensions}
\usage{
Contenus(dw, typeContenu, MethContenu = "MatDF", EmprPays = "FRA")
}
\arguments{
\item{dw}{wide data MRIO object}

\item{typeContenu}{text}

\item{MethContenu}{text}

\item{EmprPays}{text country}
}
\value{
dl data long
}
\description{
Contenus
Master function in the ecosystem
Function calculation of contents (in VA, émissions CO2, emploi) : allow footprint calculation along different formats
WARNING Needs CompoMRIO with bonus and extensions
}
\examples{
\dontrun{ List_Contenus <- Contenus(List_Interm, typeContenu = "VA", MethContenu = "MatDF", EmprPays = SelectCountry)}
}
