% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/av_fun_GVC.R
\name{ContVAdesExports}
\alias{ContVAdesExports}
\title{ContVAdesExports
Calculation of content of value addend embedded in exports for a given country and a given year
Foster-McGregor, N., et R. Stehrer (2013) : “Value added content of trade : A comprehensive approach,” Economics Letters, 120(2), 354–357.}
\usage{
ContVAdesExports(dtdl, Optdl = TRUE, annee, pays, OptUE27 = FALSE)
}
\arguments{
\item{dtdl}{data}

\item{Optdl}{binary}

\item{annee}{year}

\item{pays}{text country}

\item{OptUE27}{binary}
}
\value{
dl data long
}
\description{
ContVAdesExports
Calculation of content of value addend embedded in exports for a given country and a given year
Foster-McGregor, N., et R. Stehrer (2013) : “Value added content of trade : A comprehensive approach,” Economics Letters, 120(2), 354–357.
}
