% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/av_extensions_manager.R
\name{EmissionsProd}
\alias{EmissionsProd}
\title{EmissionsProd
 Function to calculate Inventory Emissions (coming from production)
 Be careful : [["EmiOverOuput"]] must have been introduced in our MRIO previously (with av_extend_MRIO_dw() function)}
\usage{
EmissionsProd(MRIOinterm)
}
\arguments{
\item{MRIOinterm}{MRIO object}
}
\value{
df data frame
}
\description{
EmissionsProd
 Function to calculate Inventory Emissions (coming from production)
 Be careful : [["EmiOverOuput"]] must have been introduced in our MRIO previously (with av_extend_MRIO_dw() function)
}
\examples{
\dontrun{ EmiProd_MRIO <- EmissionsProd(MRIO)}
}
