% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/av_heart_MRIO_manager.R
\name{GereInfNA}
\alias{GereInfNA}
\title{GereInfNA
 Function (very useful) to convert infinity data into NA and NA data into 0 or 1 for example}
\usage{
GereInfNA(df, impute = 0)
}
\arguments{
\item{df}{dataframe}

\item{impute}{value to inpute}
}
\value{
df data frame
}
\description{
GereInfNA
 Function (very useful) to convert infinity data into NA and NA data into 0 or 1 for example
}
\examples{
\dontrun{ if (TypeSPA == "VA") {
  f_dt <- vectDF(diag(MRIO[["VAOverOuput"]]))
  f_dt <- GereInfNA(f_dt)
}}
}
