% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/av_heart_MRIO_manager.R
\name{GetRownamesFromFirstCol}
\alias{GetRownamesFromFirstCol}
\title{GetRownamesFromFirstCol
Get Rownames From First Column}
\usage{
GetRownamesFromFirstCol(df)
}
\arguments{
\item{df}{dataframe}
}
\value{
df data frame
}
\description{
GetRownamesFromFirstCol
Get Rownames From First Column
}
