% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/av_fun_GVC.R
\name{Herfindahl}
\alias{Herfindahl}
\title{Herfindahl
Function Herfindahl (concentration indicator)}
\usage{
Herfindahl(MRIOinterm, verbose = F)
}
\arguments{
\item{MRIOinterm}{MRIO object}

\item{verbose}{binary}
}
\value{
dl data long
}
\description{
Herfindahl
Function Herfindahl (concentration indicator)
}
