% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/av_fun_GVC.R
\name{ImportedContentInVA}
\alias{ImportedContentInVA}
\title{ImportedContentInVA
Function to calculate the imported content in value added
it works with 1 country but you can loop all over the contries available in the MRIO.}
\usage{
ImportedContentInVA(pays)
}
\arguments{
\item{pays}{text country}
}
\value{
dl data long
}
\description{
ImportedContentInVA
Function to calculate the imported content in value added
it works with 1 country but you can loop all over the contries available in the MRIO.
}
