% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/av_fun_GVC.R
\name{IndicVariant_IndusCountry}
\alias{IndicVariant_IndusCountry}
\title{IndicVariant_IndusCountry
Function to calculate differences of indicators resulting from a HRM variant for instance}
\usage{
IndicVariant_IndusCountry(
  IndusREF,
  CountryREF = "FRA",
  ListCountryREF = c("FRA", "DEU", "GBR", "ESP", "ITA"),
  MRIO,
  MRIObis,
  RatioEmploiVA
)
}
\arguments{
\item{IndusREF}{text industry}

\item{CountryREF}{text country}

\item{ListCountryREF}{list}

\item{MRIO}{MRIO object}

\item{MRIObis}{MRIO object}

\item{RatioEmploiVA}{text data link}
}
\value{
dl data long
}
\description{
IndicVariant_IndusCountry
Function to calculate differences of indicators resulting from a HRM variant for instance
}
