% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/av_heart_MRIO_manager.R
\name{Info_MRIO}
\alias{Info_MRIO}
\title{Info_MRIO
Info_MRIO function: provides information on a product * country (equivalent to branch * country)
MRIO must be derived from compoMRIO
nb_top : number of crosses to display (ex: nb of countries of origin)
Output in print format and save possible.}
\usage{
Info_MRIO(MRIO, Indus_select, Country_select, nb_top = 5, verbose = T)
}
\arguments{
\item{MRIO}{MRIO object}

\item{Indus_select}{Industry selection}

\item{Country_select}{Country selection}

\item{nb_top}{Number of outputs}

\item{verbose}{verbose option}
}
\value{
text
}
\description{
Info_MRIO
Info_MRIO function: provides information on a product * country (equivalent to branch * country)
MRIO must be derived from compoMRIO
nb_top : number of crosses to display (ex: nb of countries of origin)
Output in print format and save possible.
}
\examples{
\dontrun{ Info_MRIO(MRIO, "A01", "FRA", nb_top = 5)
test <- Info_MRIO(MRIO, "C19", "FRA", nb_top = 10) / print(test)}
}
