% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/av_fun_GVC.R
\name{LinkageBwdFwd}
\alias{LinkageBwdFwd}
\title{LinkageBwdFwd
Function Backward and Forward Linkage}
\usage{
LinkageBwdFwd(base_dt, annee)
}
\arguments{
\item{base_dt}{datatable}

\item{annee}{year}
}
\value{
dl data long
}
\description{
LinkageBwdFwd
Function Backward and Forward Linkage
}
