% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/av_heart_MRIO_manager.R
\name{ListsReferential}
\alias{ListsReferential}
\title{ListsReferential
Function to return lists of countries or operations or both
You can add lists depending of new MRIO for exemple, or depending of releases}
\usage{
ListsReferential(TypeList)
}
\arguments{
\item{TypeList}{Choose the list type}
}
\value{
list
}
\description{
ListsReferential
Function to return lists of countries or operations or both
You can add lists depending of new MRIO for exemple, or depending of releases
}
\examples{
\dontrun{ if (TypeList == "BRFD_FIG_WIOD_LRWIOD_ICIO") {
  List_out <- append(ListsReferential("BR_FIG_WIOD_LRWIOD_ICIO"), ListsReferential("FD_FIG_WIOD_LRWIOD_ICIO"))
}
return(List_out)}
}
