% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/av_fun_GVC.R
\name{MadeIn_byOrigin}
\alias{MadeIn_byOrigin}
\title{MadeIn_byOrigin
Calculate MAdeIn by country of origin, for analytics}
\usage{
MadeIn_byOrigin(
  RESULT_MADEINs_byCountry,
  SelectCountries = c("FRA", "DEU", "GBR", "ESP", "ITA", "USA", "CHN"),
  MadeInOf = "FRA",
  OptSaveRDS = FALSE
)
}
\arguments{
\item{RESULT_MADEINs_byCountry}{datatable}

\item{SelectCountries}{list}

\item{MadeInOf}{text country}

\item{OptSaveRDS}{binary}
}
\value{
dl data long
}
\description{
MadeIn_byOrigin
Calculate MAdeIn by country of origin, for analytics
}
