% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/av_fun_GVC.R
\name{MadeIn_Manuf}
\alias{MadeIn_Manuf}
\title{MadeIn_Manuf
Calculation of manuf's made-in by agreggation (2 classifications)}
\usage{
MadeIn_Manuf(
  resMadeIn_LRWIOD,
  resMadeIn_WIOD,
  resMadeIn_FIGARO,
  resMadeIn_ICIO,
  OptSaveRDS = FALSE
)
}
\arguments{
\item{resMadeIn_LRWIOD}{datatable}

\item{resMadeIn_WIOD}{datatable}

\item{resMadeIn_FIGARO}{datatable}

\item{resMadeIn_ICIO}{datatable}

\item{OptSaveRDS}{binary}
}
\value{
dl data long
}
\description{
MadeIn_Manuf
Calculation of manuf's made-in by agreggation (2 classifications)
}
