% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/av_fun_GVC.R
\name{MadeIn_Retropolation}
\alias{MadeIn_Retropolation}
\title{MadeIn_Retropolation
Retropolation of made-in to fit long series inter-MRIO
You can manage the MRIO list}
\usage{
MadeIn_Retropolation(
  resMadeIn_manuf,
  MadeIn_levels,
  resMadeIn_LRWIOD,
  resMadeIn_WIOD,
  resMadeIn_FIGARO,
  resMadeIn_ICIO,
  OptSaveRDS = FALSE
)
}
\arguments{
\item{resMadeIn_manuf}{datatable}

\item{MadeIn_levels}{datatable}

\item{resMadeIn_LRWIOD}{datatable}

\item{resMadeIn_WIOD}{datatable}

\item{resMadeIn_FIGARO}{datatable}

\item{resMadeIn_ICIO}{datatable}

\item{OptSaveRDS}{binary}
}
\value{
dl data long
}
\description{
MadeIn_Retropolation
Retropolation of made-in to fit long series inter-MRIO
You can manage the MRIO list
}
