% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/av_fun_GVC.R
\name{MadeIn_Retropolation_UE27}
\alias{MadeIn_Retropolation_UE27}
\title{MadeIn_Retropolation_UE27
Retropolation of made-in to fit long series inter-MRIO}
\usage{
MadeIn_Retropolation_UE27(
  resMadeInUE27_LRWIOD,
  resMadeInUE27_WIOD,
  resMadeInUE27_FIGARO,
  resMadeInUE27_ICIO,
  OptSaveRDS = FALSE
)
}
\arguments{
\item{resMadeInUE27_LRWIOD}{datatable}

\item{resMadeInUE27_WIOD}{datatable}

\item{resMadeInUE27_FIGARO}{datatable}

\item{resMadeInUE27_ICIO}{datatable}

\item{OptSaveRDS}{binary}
}
\value{
dl data long
}
\description{
MadeIn_Retropolation_UE27
Retropolation of made-in to fit long series inter-MRIO
}
