% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Mult2_rcpp3}
\alias{Mult2_rcpp3}
\title{Matrix Multiplication : 2 components}
\usage{
Mult2_rcpp3(matA, matB)
}
\arguments{
\item{matA}{matrix}

\item{matB}{matrix}
}
\description{
This function returns the matrix multiplication of 2 matrix
}
