% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/av_heart_MRIO_manager.R
\name{PasteN}
\alias{PasteN}
\title{PasteN
Function to concatenate the amounts (in value) of several columns of a DT (in list form)}
\usage{
PasteN(dt, column_ref)
}
\arguments{
\item{dt}{datatable}

\item{column_ref}{column number}
}
\value{
text
}
\description{
PasteN
Function to concatenate the amounts (in value) of several columns of a DT (in list form)
}
\examples{
\dontrun{ P2 <- PasteN(DF_TOT_EXP_topN_w, c("Col_Country", "Col_Indus"))}
}
