% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/av_fun_GVC.R
\name{RatioNoguera}
\alias{RatioNoguera}
\title{RatioNoguera
Calculation of Noguera indicator}
\usage{
RatioNoguera(BaseResult, nomMRIO)
}
\arguments{
\item{BaseResult}{datatable}

\item{nomMRIO}{text name MRIO}
}
\value{
dl data long
}
\description{
RatioNoguera
Calculation of Noguera indicator
}
