% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/av_heart_MRIO_manager.R
\name{ReqSum}
\alias{ReqSum}
\title{ReqSum
Query function to calculate a sum from a list of dimensions to be summed
ListDimASommer is the normalized list of dimensions on wich you wants to sum up.
OptStruct keeps initial format of data bur compute structure on the given list of dimensions (ListDimASommer)}
\usage{
ReqSum(dl, ListDimASommer, OptStruct = FALSE)
}
\arguments{
\item{dl}{data_long format database}

\item{ListDimASommer}{list of dimensions}

\item{OptStruct}{option for structure results}
}
\value{
dl data long
}
\description{
ReqSum
Query function to calculate a sum from a list of dimensions to be summed
ListDimASommer is the normalized list of dimensions on wich you wants to sum up.
OptStruct keeps initial format of data bur compute structure on the given list of dimensions (ListDimASommer)
}
\examples{
\dontrun{ Cadre <- Base_depart[["DF"]]
Cadre_Tot <- ReqSum(Cadre, "Col_Indus") # Sum components of final demand}
}
