% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/av_heart_MRIO_manager.R
\name{rndN}
\alias{rndN}
\title{rndN
Rounding function, default 0 decimal places}
\usage{
rndN(dataa, rnd = 0)
}
\arguments{
\item{dataa}{data to round}

\item{rnd}{round decimal option}
}
\value{
rounded value
}
\description{
rndN
Rounding function, default 0 decimal places
}
