% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/av_heart_MRIO_manager.R
\name{SoldExtPays}
\alias{SoldExtPays}
\title{SoldExtPays
Function to calculate external balance for all countries of a MRIO}
\usage{
SoldExtPays(MRIOinterm)
}
\arguments{
\item{MRIOinterm}{MRIO object}
}
\value{
df data frame
}
\description{
SoldExtPays
Function to calculate external balance for all countries of a MRIO
}
\examples{
\dontrun{ SoldeExt <- SoldExtPays(MRIO)}
}
