% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/av_heart_MRIO_manager.R
\name{SommeDFenP3_S14}
\alias{SommeDFenP3_S14}
\title{SommeDFenP3_S14
Function which sums up all final demand and stores it in P3_14
Caution: potentially biases detailed interpretation of MRIO components
On the other hand, remains compatible with all functions and avoids the weird things of P5M (stock variatins that make HRM jump on C22 in particular)
Only works with FIGARO at this time, needs more lists if you wants to expand}
\usage{
SommeDFenP3_S14(base_dl)
}
\arguments{
\item{base_dl}{long database}
}
\value{
dl data long
}
\description{
SommeDFenP3_S14
Function which sums up all final demand and stores it in P3_14
Caution: potentially biases detailed interpretation of MRIO components
On the other hand, remains compatible with all functions and avoids the weird things of P5M (stock variatins that make HRM jump on C22 in particular)
Only works with FIGARO at this time, needs more lists if you wants to expand
}
\examples{
\dontrun{ if (OptSommeDFenP3_S14 == TRUE) {
  Base_init <- SommeDFenP3_S14(data.table::copy(dl))
}}
}
