% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/av_heart_MRIO_manager.R
\name{SplitPRBR}
\alias{SplitPRBR}
\title{SplitPRBR
Split of PR and BR into four parts or only PR or only BR
Please note: country names, branch names and product names must not contain underscores.}
\usage{
SplitPRBR(dl)
}
\arguments{
\item{dl}{data long, like normalized database}
}
\value{
dl data long
}
\description{
SplitPRBR
Split of PR and BR into four parts or only PR or only BR
Please note: country names, branch names and product names must not contain underscores.
}
\examples{
\dontrun{ test <- SplitPRBR(DT) }
}
