% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/av_heart_MRIO_manager.R
\name{vectDF}
\alias{vectDF}
\title{vectDF
 Function to convert a vector into dataframe}
\usage{
vectDF(vect)
}
\arguments{
\item{vect}{vector}
}
\value{
df data frame
}
\description{
vectDF
 Function to convert a vector into dataframe
}
\examples{
\dontrun{ MyDF <- vectDF(DF[, 1])}
}
